+++ common/src/main/java/lol/zanspace/unloadedactivity/mixin/chunk/precipitationTicks/LayeredCauldronMixin.java
@@ -67,13 +67,13 @@ public abstract class LayeredCauldronMixin extends AbstractCauldronBlock {
     }
 
     @Override
-    public void simulatePrecTicks(BlockState state, ServerLevel level, BlockPos pos, long timeInWeather, long timePassed, Biome.Precipitation precipitation, double precipitationPickChance) {
+    public void simulatePrecTicks(BlockState state, ServerLevel level, BlockPos pos, long timeInWeather, long timePassed, Biome.Precipitation precipitation, float precipitationPickChance) {
 
         int maxCauldronLevel = LayeredCauldronBlock.MAX_FILL_LEVEL;
         int currentCauldronLevel = state.getValue(LayeredCauldronBlock.LEVEL);
         int levelDifference = maxCauldronLevel-currentCauldronLevel;
 
-        double totalOdds = precipitationPickChance*getFillOdds(precipitation);
+        float totalOdds = precipitationPickChance*getFillOdds(precipitation);
 
         int fill = Utils.getOccurrencesBinomial(timeInWeather, totalOdds, levelDifference, level.random);
